Remove unnecessary ForEach's in LiveContext.buildChildren
This brings the time of a particular view update from 700ms to 300ms.
Something about ForEach or the additional closures meant we were
spending an incredible amount of time dealing with Swift type metadata
at runtime (swift_getTypeByMangledName accounted for about 575ms of the
time). This brings it down to 200ms, which is still a lot but is
substantially better.