Use regular, non-testable imports for rendering tests
When one of our (non-generic) view types clashes with a SwiftUI one
(e.g., Spacer), SwiftUI seems to stop propagating the environment (and
possibly other parts of the attribute graph). This caused the spacer
test to crash, since it couldn't access the element (which is via the
environment) to look up an attribute value.