Fix retain cycle in LiveViewCoordinator
`LiveViewCoordinator` is initialised from `LiveSessionCoordinator`, which
gets passed in as a parameter to `init` and then retained by the view
coordinator. This causes a retain cycle.

Additionally, this is also retained by the `handleEvent` as `session`
was used directly in the closure, which then retained it.

This commit fixes both by using a weak reference from the view coordinator
to the session coordinator, and accessing the session via the weak `self`
reference.